[book] route done

+++ src/app/modules/books/book.service.ts
@@ -1,15 +1,28 @@
 import httpStatus from 'http-status';
-import { SortOrder } from 'mongoose';
+import { SortOrder, Types } from 'mongoose';
 import ApiError from '../../../errors/ApiError';
 import { paginationHelpers } from '../../../helpers/pagination.helpers';
 import { IGenericResponse } from '../../../interfaces/common';
 import { IPaginationOptions } from '../../../interfaces/pagination';
 import { bookSearchableFields } from './book.constants';
-import { IBook, IBookFilters } from './book.interface';
+import { IBook, IBookFilters, IReview } from './book.interface';
 import { BookModel } from './book.model';
 
-const addBook = async (bookData: IBook): Promise<IBook | null> => {
-  return await BookModel.create(bookData);
+const addBook = async (
+  userId: Types.ObjectId,
+  bookData: IBook
+): Promise<IBook | null> => {
+  const isExist = await BookModel.findOne({ title: bookData.title });
+  if (isExist) {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      'Book already exists with this title'
+    );
+  }
+
+  return (await BookModel.create({ ...bookData, uploadedBy: userId })).populate(
+    'uploadedBy'
+  );
 };
 
 const getAllBook = async (
@@ -55,11 +68,20 @@ const getAllBook = async (
     andConditions.length > 0 ? { $and: andConditions } : {};
 
   const result = await BookModel.find(whereConditions)
+    .populate('uploadedBy')
+    .populate({
+      path: 'reviews',
+      populate: [
+        {
+          path: 'user',
+        },
+      ],
+    })
     .sort(sortConditions)
     .skip(skip)
     .limit(limit)
     .lean();
-  const total = await BookModel.countDocuments(whereConditions);
+  const total = await BookModel.countDocuments(whereConditions).lean();
   return {
     meta: {
       page,
@@ -75,15 +97,43 @@ const getSingleBook = async (id: string): Promise<IBook | null> => {
   if (!isExist) {
     throw new ApiError(httpStatus.NOT_FOUND, 'Book do not exist!');
   }
-  return await BookModel.findById(id);
+  return await BookModel.findById(id)
+    .populate('uploadedBy')
+    .populate({
+      path: 'reviews',
+      populate: [
+        {
+          path: 'user',
+        },
+      ],
+    });
 };
 
-const deleteBook = async (id: string): Promise<IBook | null> => {
-  const isExist = await BookModel.findById(id);
+const deleteBook = async (
+  userId: Types.ObjectId,
+  id: string
+): Promise<IBook | null> => {
+  const isExist = await BookModel.findById(id).populate('uploadedBy');
   if (!isExist) {
     throw new ApiError(httpStatus.NOT_FOUND, 'Book do not exist!');
   }
-  return await BookModel.findByIdAndDelete(id);
+
+  if (isExist.uploadedBy?._id != userId) {
+    throw new ApiError(
+      httpStatus.UNAUTHORIZED,
+      'You do not have permission to delete this!'
+    );
+  }
+  return await BookModel.findByIdAndDelete(id)
+    .populate('uploadedBy')
+    .populate({
+      path: 'reviews',
+      populate: [
+        {
+          path: 'user',
+        },
+      ],
+    });
 };
 
 const updateBook = async (
@@ -100,7 +150,52 @@ const updateBook = async (
 
   const result = await BookModel.findByIdAndUpdate(id, updatedBookData, {
     new: true,
-  });
+  })
+    .populate('uploadedBy')
+    .populate({
+      path: 'reviews',
+      populate: [
+        {
+          path: 'user',
+        },
+      ],
+    });
+  return result;
+};
+const addReview = async (
+  userId: Types.ObjectId,
+  bookId: string,
+  payload: Partial<IReview>
+): Promise<IBook | null> => {
+  const isExist = await BookModel.findById(bookId);
+
+  if (!isExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Book do not exist!');
+  }
+
+  const updatedBookData: Partial<IReview> = { ...payload };
+  updatedBookData.user = userId;
+
+  const result = await BookModel.findByIdAndUpdate(
+    bookId,
+    {
+      $push: {
+        reviews: updatedBookData,
+      },
+    },
+    {
+      new: true,
+    }
+  )
+    .populate('uploadedBy')
+    .populate({
+      path: 'reviews',
+      populate: [
+        {
+          path: 'user',
+        },
+      ],
+    });
   return result;
 };
 
@@ -110,4 +205,5 @@ export const BookService = {
   getSingleBook,
   deleteBook,
   updateBook,
+  addReview,
 };
